fix: use uniquery instead
