+++ ogi/package.json
@@ -23,7 +23,6 @@
     "@kodadot1/minipfs": "^0.4.2-rc.0",
     "@kodadot1/static": "^0.0.3",
     "@kodadot1/uniquery": "^0.5.1-rc.0",
-    "@polkadot/util": "^12.5.1",
-    "nuxt-graphql-client": "^0.2.31"
+    "@polkadot/util": "^12.5.1"
   }
 }
