--- ogi/queries/nftById.ahp.gql
@@ -1,12 +1,0 @@
-query nftById_ahp($id: String!) {
-  nftEntityById(id: $id) {
-    id
-    name
-    price
-    meta {
-      id
-      image
-      description
-    }
-  }
-}
