--- ogi/queries/nftById.bsx.gql
@@ -1,12 +1,0 @@
-query nftById_bsx($id: String!) {
-  nftEntityById(id: $id) {
-    id
-    name
-    price
-    meta {
-      id
-      image
-      description
-    }
-  }
-}
