--- ogi/queries/nftById.gql
@@ -1,12 +1,0 @@
-query nftById_ahk($id: String!) {
-  nftEntityById(id: $id) {
-    id
-    name
-    price
-    meta {
-      id
-      image
-      description
-    }
-  }
-}
