--- ogi/queries/nftById.ksm.gql
@@ -1,12 +1,0 @@
-query nftById_ksm($id: String!) {
-  nftEntityById(id: $id) {
-    id
-    name
-    price
-    meta {
-      id
-      image
-      description
-    }
-  }
-}
