--- ogi/queries/nftById.rmrk.gql
@@ -1,12 +1,0 @@
-query nftById_rmrk($id: String!) {
-  nftEntityById(id: $id) {
-    id
-    name
-    price
-    meta {
-      id
-      image
-      description
-    }
-  }
-}
